requirements.txt
playwright
python-dotenv
requests
playwright